requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.3
python-multipart>=0.0.6
langchain>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.1
motor>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
email-validator>=2.1.0
gunicorn>=21.2.0
